HW_5 - Fixed final HW_5 Added base_page.py Added locators.py Added registration_page.py Added welcome_page.py Added __init__.py Changed locator for WELCOME text
