--- Week_5/HW_5/allure_report/2d79ca5a-f472-4488-a377-67339b2b4198-result.json
@@ -1,1 +1,0 @@
-{"name": "test_user_successfully_register", "status": "passed", "attachments": [{"name": "stdout", "source": "7d2db06b-87a5-49c2-bd2e-1f5cb74fd648-attachment.txt", "type": "text/plain"}], "start": 1594845282679, "stop": 1594845288778, "uuid": "1fc9a657-9a8a-4040-9610-c576d4119080", "historyId": "56a8912d72b8cc9e16ed32265ddb30d8", "testCaseId": "43411d151b12c89ed2564dcfbbf03970", "fullName": "Week_5.HW_5.test_make_report.TestRegistrationNewUser#test_user_successfully_register", "labels": [{"name": "parentSuite", "value": "Week_5.HW_5"}, {"name": "suite", "value": "test_make_report"}, {"name": "subSuite", "value": "TestRegistrationNewUser"}, {"name": "host", "value": "PRGLAPFLLZGH2"}, {"name": "thread", "value": "19732-MainThread"}, {"name": "framework", "value": "pytest"}, {"name": "language", "value": "cpython3"}, {"name": "package", "value": "Week_5.HW_5.test_make_report"}]}
