--- Week_5/HW_5/allure_report/7d2db06b-87a5-49c2-bd2e-1f5cb74fd648-attachment.txt
@@ -1,13 +1,0 @@
-
-start Chrome browser for test..
-1594845282.6790001@fakemail.com
-Summercamp2020
-we should be on MAIN page
-we should be on Registration page
-sleep 2 seconds
-email entered
-password entered
-confirmation password entered
-now will wait 3 seconds
-
-quit browser..
