--- Week_5/HW_5/allure_report/a684dfb0-ad18-4405-ae1d-cd0d33baad31-container.json
@@ -1,1 +1,0 @@
-{"uuid": "70849851-ef5e-4877-a786-e1ff589fd421", "children": ["1fc9a657-9a8a-4040-9610-c576d4119080"], "befores": [{"name": "browser", "status": "passed", "start": 1594845280540, "stop": 1594845282676}], "afters": [{"name": "browser::0", "status": "passed", "start": 1594845288779, "stop": 1594845291984}], "start": 1594845280540, "stop": 1594845291984}
