+++ Week_5/HW_5/base_page.py
@@ -1,0 +1,23 @@
+from selenium.common.exceptions import NoSuchElementException
+from .locators import BasePageLocators
+
+
+class BasePage():
+    def __init__(self, browser, url, timeout=5):
+        self.browser = browser
+        self.url = url
+        self.browser.implicitly_wait(timeout)
+
+    def go_to_login_page(self):
+        link = self.browser.find_element(*BasePageLocators.LOGIN_LINK)
+        link.click()
+
+    def open(self):
+        self.browser.get(self.url)
+
+    def is_element_present(self, how, what):
+        try:
+            self.browser.find_element(how, what)
+        except NoSuchElementException:
+            return False
+        return True
